Intorducing vtpmd service to bring fully virtualized TPM to the VMs
This commit brings SWTPM in EVE, in order to emulate Virtual TPM in
the VMS. To achive this we create a SWTPM instance per VM. The SWTPM
instance is configured to use a Unix Domain Socket as a communication line,
after passing the socket path to the QEMU virtual TPM configuration,
it automatically creates a virtual TPM device for the VM which is accessible
like a normal TPM device under /dev/tpm* .

SWTPM saves and loads the TPM state on/from the disk, so at the next VM
boot all the TPM keys, TPM NVRAM data, etc are present. In addition SWTPM
is configured to encrypt the TPM state files using a 256-bit AES key. We
use the same key as vault key, this key stored in TPM and access to it is
authorized using a PCR policy, this makes it easier to manage since if
we use a separate key for SWTPM states, there is one more key to
backup/restore from the controller.

Signed-off-by: Shahriyar Jalayeri <shahriyar@zededa.com>